QOS: Fixed QosInfo Success Rate int division giving only 0.0 or 1.0

+++ openschema-android/OpenSchemaSDK/src/main/java/io/openschema/mma/utils/QosInfo.java
@@ -60,7 +60,7 @@ public class QosInfo {
 
     private float calculateSuccessRate() {
         if(mRttValues.isEmpty()) return 0;
-        return (mRttValues.size() - mTotalFailedRequests)/mRttValues.size();
+        return (float)(mRttValues.size() - mTotalFailedRequests)/mRttValues.size();
     }
 
     public void cleanData() {
